old-tests/submission/Microsoft/semantic/section.html

<!DOCTYPE html>
<html><head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type">
  <title>verify SECTION	is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 <link href="http://www.w3.org/TR/html5/the-section-element.html#the-section-element" rel="help">
</head>
<body>
<div id="log"></div>
<section id="section">
 <h1>H1 in an article and secion</h1>
 <p>paragraph</p>
</section>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("section").toString() )
 })
</script>


</body></html>
